+++ src/app/components/admin/admin-events/admin-events.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { EventsService } from 'src/app/services/events.service';
+import { DialogService } from 'src/app/services/dialog.service';
 
 @Component({
   selector: 'app-admin-events',
@@ -13,7 +14,7 @@ export class AdminEventsComponent implements OnInit {
   displayedColumns: string[] = ['eventName', 'eventDescription', 'eventImgPath', 'actions'];
   deleted: boolean = false;
 
-  constructor(private eventsService:EventsService) { }
+  constructor(private eventsService:EventsService, private dialogService: DialogService) { }
 
   ngOnInit(): void {
     this.eventsService.getEvents().subscribe(data=>{
@@ -24,11 +25,15 @@ export class AdminEventsComponent implements OnInit {
   }
 
   delete(id){
-    this.eventsService.deleteEvent(id).subscribe(data=>{
-      console.log('deleted', data);
-      this.deleted = true;
-      this.ngOnInit();
-    })
+    this.dialogService.openConfirmDialog("Are you sure to delete this blog?")
+      .afterClosed().subscribe(res =>{
+        this.eventsService.deleteEvent(id).subscribe(data=>{
+          console.log('deleted', data);
+          this.deleted = true;
+          this.ngOnInit();
+        })
+      });
+    
   }
 
 }
